requirements.txt
cython==3.0.0
datasets==3.1.0
dill==0.3.7
einops==0.6.1
ffmpeg-python==0.2.0
gdown==5.2.0
gitdb==4.0.10
gitpython==3.1.32
gputil==1.4.0
h5py==3.9.0
hydra-core==1.3.2
hydra-joblib-launcher
hydra-submitit-launcher==1.2.0
imageio==2.31.1
imageio-ffmpeg==0.4.8
imgviz==1.7.4
ipdb==0.13.13
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
json5==0.9.14
matplotlib==3.7.2
matplotlib-inline==0.1.6
networkx==3.1
ninja==1.11.1  # for quicker tcnn build
numpy==1.22.4
numba==0.60.0
onnx==1.14.0
onnxruntime==1.15.1
open3d==0.16.0
opencv-python
pandas==2.0.3
# pre-commit==4.0.1
pycocotools==2.0.7
pyglet==1.5.27
pyopengl==3.1.0
pyvirtualdisplay==3.0
pyvista==0.41.1
pyyaml==6.0.1
rtree==1.0.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
shapely==2.0.1
snakeviz==2.2.0
termcolor==2.3.0
timm==0.9.5
trimesh==3.23.3
urdf-parser-py==0.0.4
wandb==0.15.8
yappi==1.4.0
